Fixed major bug
Fixed game freezing after using things from the mod.
