+++ DevilFruitMod/BombBombFruit/BombHuman.cs
@@ -2,6 +2,9 @@ using System;
 using Terraria;
 using Terraria.GameInput;
 using Terraria.ID;
+using Terraria.ModLoader;
+using DevilFruitMod.Util;
+using Microsoft.Xna.Framework;
 
 namespace DevilFruitMod.BombBombFruit
 {
@@ -15,11 +18,7 @@ namespace DevilFruitMod.BombBombFruit
             if (player.GetModPlayer<DevilFruitUser>().eatenDevilFruit == 4 && player.HeldItem.type == ItemID.None && !(player.wet && !(player.honeyWet || player.lavaWet)) && (!Equals(DevilFruitMod.UsePowers1Hotkey.GetAssignedKeys(InputMode.Keyboard)[0], "Mouse1") || (Main.hasFocus && !Main.LocalPlayer.mouseInterface && !Main.drawingPlayerChat && !Main.editSign && !Main.editChest && !Main.blockInput && !Main.mapFullscreen && !Main.HoveringOverAnNPC && Main.LocalPlayer.talkNPC == -1)))
             {
                 //Getting the shooting trajectory
-                float clickX = (int)(Main.mouseX) - Main.screenWidth / 2;
-                float clickY = (int)(Main.mouseY) - Main.screenHeight / 2;
-                float magnitude = (float)Math.Sqrt(clickX * clickX + clickY * clickY);
-                float directionX = 10 * clickX / magnitude;
-                float directionY = 10 * clickY / magnitude;
+                Vector2 dir = TMath.CalculateTrajectory();
 
                 if (DevilFruitMod.MiscHotkey.JustPressed)
                 {
@@ -28,17 +27,17 @@ namespace DevilFruitMod.BombBombFruit
 
                 if (DevilFruitMod.UsePowers1Hotkey.JustPressed)
                 {
-                    BombBombPowers(directionX, directionY, 0);
+                    BombBombPowers(dir.X, dir.Y, 0);
                 }
 
                 if (DevilFruitMod.UsePowers2Hotkey.JustPressed)
                 {
-                    BombBombPowers(directionX, directionY, 1);
+                    BombBombPowers(dir.X, dir.Y, 1);
                 }
 
                 if (DevilFruitMod.UsePowers3Hotkey.JustPressed)
                 {
-                    BombBombPowers(directionX, directionY, 2);
+                    BombBombPowers(dir.X, dir.Y, 2);
                 }
 
                 /*
@@ -47,6 +46,14 @@ namespace DevilFruitMod.BombBombFruit
 
                 }
                 */
+            }
+        }
+
+        public override void PreUpdate()
+        {
+            if (player.GetModPlayer<DevilFruitUser>().eatenDevilFruit > 0 && player.wet && !(player.honeyWet || player.lavaWet))
+            {
+                player.AddBuff(ModContent.BuffType<Buffs.waterStun>(), 60, true);
             }
         }
