+++ DevilFruitMod/DFTemplateDFTemplateFruit/DFTemplateHuman.cs
@@ -1,3 +1,4 @@
+using DevilFruitMod.Util;
 using System;
 using Terraria;
 using Terraria.GameInput;
@@ -29,11 +30,7 @@ namespace DevilFruitMod.DFTemplateDFTemplateFruit
             if (player.GetModPlayer<DevilFruitUser>().eatenDevilFruit == 99 && player.HeldItem.type == ItemID.None && !(player.wet && !(player.honeyWet || player.lavaWet)) && (!Equals(DevilFruitMod.UsePowers1Hotkey.GetAssignedKeys(InputMode.Keyboard)[0], "Mouse1") || (Main.hasFocus && !Main.LocalPlayer.mouseInterface && !Main.drawingPlayerChat && !Main.editSign && !Main.editChest && !Main.blockInput && !Main.mapFullscreen && !Main.HoveringOverAnNPC && Main.LocalPlayer.talkNPC == -1)))
             {
                 //Getting the shooting trajectory
-                float clickX = (int)(Main.mouseX) - Main.screenWidth / 2;
-                float clickY = (int)(Main.mouseY) - Main.screenHeight / 2;
-                float magnitude = (float)Math.Sqrt(clickX * clickX + clickY * clickY);
-                float directionX = 10 * clickX / magnitude;
-                float directionY = 10 * clickY / magnitude;
+                Vector2 dir = TMath.CalculateTrajectory();
 
                 if (DevilFruitMod.MiscHotkey.JustPressed)
                 {
@@ -42,17 +39,17 @@ namespace DevilFruitMod.DFTemplateDFTemplateFruit
 
                 if (DevilFruitMod.UsePowers1Hotkey.JustPressed)
                 {
-                    DFTemplateDFTemplatePowers(directionX, directionY, 0);
+                    DFTemplateDFTemplatePowers(dir.X, dir.Y, 0);
                 }
 
                 if (DevilFruitMod.UsePowers2Hotkey.JustPressed)
                 {
-                    DFTemplateDFTemplatePowers(directionX, directionY, 1);
+                    DFTemplateDFTemplatePowers(dir.X, dir.Y, 1);
                 }
 
                 if (DevilFruitMod.UsePowers3Hotkey.JustPressed)
                 {
-                    DFTemplateDFTemplatePowers(directionX, directionY, 2);
+                    DFTemplateDFTemplatePowers(dir.X, dir.Y, 2);
                 }
 
                 /*
@@ -61,6 +58,14 @@ namespace DevilFruitMod.DFTemplateDFTemplateFruit
 
                 }
                 */
+            }
+        }
+
+        public override void PreUpdate()
+        {
+            if (player.GetModPlayer<DevilFruitUser>().eatenDevilFruit > 0 && player.wet && !(player.honeyWet || player.lavaWet))
+            {
+                player.AddBuff(ModContent.BuffType<Buffs.waterStun>(), 60, true);
             }
         }
